feat(tree class modified)
The tree class now has a constant variable called sortedArray which uses the later called sortAndRemoveDuplicates function. Then, for the root, it also uses a later called function called buildTree, using the sortedArray as one of the parameters. The sortAndRemoveDuplicates function will take an array as argument, sort it in ascendant order using the sort method, then using that sorted array to filter it, creating a new array that contains only the unique values.