feat(buildtree property)
The buildtree function will take an array, a start and a ending as parameters. The first conditional will be the base case, checking if the start is greater than the ending, returing null. Then a mid and root variables are created, the first one containing the sum of start and end then divided by two, with parseInt method applied to the result, and the root creating a new node object with the array at the mid index. For both leftChild and rightChild uses the buildTree recursively, using the array, start and mid -1 for the left as parameters, and array, mid +1 and end for the right. Then the root element is returned.