refactor(Tree class properties refactored)
The Tree class now only contains one property called root that invokes the buildTree with the given array as parameter. sortAndRemoveDuplicates now uses another method to sort the array, by using the Set method with the array and putting it into an array directly to then sort it using a given function as example. The buildTree sorts the parameter array, then sets a base case if the array length is equal to 0, creates the mid part of the array dividing the array lenght by 2 and parseInt that, then creates the root of the tree using the Node constructor giving the mid of the array as data, 0 to mid as the left array and and mid + 1 as the right array. Finally, root is returned.