feat(insert property)
The insert property uses two parameters, the first one the given value of the node to be inserted inside the tree, the second one being the root node. The first conditional checks if wether the root exists or not, if not, a new node is created with the given value, being placed as the root node. Then, another conditional checking if the root.data is less than the given value with two ternary operators, the first one for the truthy case which assings the calling of insert with the given value and the root.right as second parameter to the root.right of the current root node, doing the opposite if not true.