feat(detele property)
The property takes two arguments, the first being the value of the node to be deleted and the second one being the current root of the BTS or the current starting node. Then it checks with a base case if there is a nood, if not, returns the root used in the parameter. A new conditional appears, checkig if wether the root.data is lesser than the value, using the function recursively on the rigth tree of the node. If the opposite happens, then the function is used recursively on the left tree. If either of these two cases are true, it means that you find the exact node to be deleted, and then other conditionals are used to handle how the children nodes are rearrange.