feat(minValue property)
The functions uses a node as a parameter and creates a min value variable with the node.data. Then it creates a loop using while, checking if the node.left exists, so as long as that is true, the new minValue will be the node.left.data, and node itself is updated to be the current node.left, until it reaches a node that does not have a node.left, returning the last node.data and then finding the min value of the original node.right.