feat(find property)
The functions uses a value and the current root node as parameters, the first checking is if the root is null which will return effectively the root. Then it checks if the root.data and the value provided are not equal, meaning that the node wanted is not the curent node, and in that case will then check if the value is greater or lesser than the node.data, calling the find function recursively on the right if the number is greater, or on the left if the opposite happens. Finally, the root is returned as the number was found.