feat(inoder property)
The inorder function will take three parameters, the node being the root node, a callback and a result array. The first two base cases handle the case of not having a root, which will return an empty array or having a node with a null value, which returns and ends the function. After that, the function is recursively called, using the left sub-tree or node.left, a callback and a result. If a callback exists, then the callback is used on the node, if not, the node.data is pushed to the result array. After that, the same inorder functions is recursively called on the right node. At the end, if there is a result array, it is returned.