feat(depth method)
The depth function takes three parameters, the node to which the depth is going to be calculated, the default root and the level starting at 0. The base case will test if wether there is no node, so null is returned, if the root is equal to null then 0 is returned and if the root.data is equal to the node.data, in that case the level is returned. If neither of those conditions are met, a new varaible called count is created with a recursive use of depth using the root.left as one of the parameters and incresing the level in one. If count is still 0, the function continues, this time returning the recursive call of the depth function but with the root.right, always increasing the level in one.