feat(isbalanced method)
The function will use the default root as a node. THe base case checks if wether the node has null as value, returning true, as the three will be balance. THen a heightDiff variable will store the result of math.abs over height(node.left - height(node.right). Returning true if the heightDiff value is less or equal to 1, and if the same result is obtained by recursively calling the isBalance function on the node.left and node.right sides and all its children nodes.