feat(rebalance method)
The rebalance function has a base case checkinf if the root has null as value, returning if true. If not, a variable called sorted will store an array using the ...new Set over the inorder function and sorting method. Finally, the root of the tree will have a value of buildTree(sorted).